refactor(workbench): extract composer dictation FSM into useVoiceDictation
Host still owns voice settings and send(); capture, STT, draft insert, and live-voice overlay open state live in the hook so dictation changes do not require opening AppWorkbench internals.
